remove: `libloading` dependency
