+++ autd3-link-twincat/Cargo.toml
@@ -15,13 +15,12 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
 default = ["local"]
-local = ["libloading"]
+local = []
 remote = ["cc"]
 all = ["local", "remote"]
 
 [dependencies]
 autd3-core = { workspace = true, features = ["std", "link"] }
-libloading = { version = "0.8.6", optional = true }
 
 [build-dependencies]
 cc = { version = "1.2.24", optional = true }
